database issue is fixed as per last commit.
Calendar displays correct dates now.

Calendar navigates to List of tasks when a date is selected. (Even if no dates are selected)

Picking a task loads that task in edit task view.

You can edit or delete the task. You are redirected to the calendar. Redirecting to the task list resulted in a crash due to not being able run sql queries thanks to intents not carrying values, or database queries not getting values due to the change. This will need to be fixed further.